+++ src/navbar/types.ts
@@ -87,6 +87,7 @@ export interface NavbarActionButtonCtaProps {
 
 /** Props for the internal `"component"` type action button wrapper. */
 interface NavbarButtonComponentProps {
+    /** The custom component to be rendered */
     render: JSX.Element;
 }
 
@@ -194,6 +195,7 @@ export type DrawerDismissalMethod =
     | "brand-click";
 
 // TODO: Make this to be an internal type
+// @storybookSkipProps
 /**
  * Props for the internal mobile navigation drawer panel.
  */
@@ -232,6 +234,7 @@ export interface NavbarDrawerProps extends NavbarSharedProps {
  */
 export type BrandType = "primary" | "secondary";
 
+// @storybookSkipProps
 /**
  * Imperative methods exposed by the `Navbar` component via `ref`.
  */
@@ -243,6 +246,7 @@ export type NavbarDrawerApi = {
     dismissDrawer: () => void;
 };
 
+// @storybookSkipProps
 /**
  * Imperative handle returned by `Navbar` via `ref`.
  *
