+++ stories/language-switcher/language-switcher.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta } from "@storybook/addon-docs/blocks";
 import * as LanguageSwitcherStories from "./language-switcher.stories";
-import { PropsTable } from "./props-table";
+import { AutoArgTypesTabs } from "../storybook-common";
 
 <Meta of={LanguageSwitcherStories} />
 
@@ -54,4 +54,4 @@ import { LanguageSwitcher } from "@lifesg/react-design-system/language-switcher"
 
 ## Component API
 
-<PropsTable />
+<AutoArgTypesTabs of={LanguageSwitcherStories} />
