+++ stories/link-list/link-list.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta } from "@storybook/addon-docs/blocks";
 import * as LinkListStories from "./link-list.stories";
-import { PropsTable } from "./props-table";
+import { AutoArgTypesTabs } from "../storybook-common";
 
 <Meta of={LinkListStories} />
 
@@ -42,4 +42,4 @@ You can customise the `View more` and `View less` labels.
 
 ## Component API
 
-<PropsTable />
+<AutoArgTypesTabs of={LinkListStories} />
