+++ stories/local-nav/local-nav.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta } from "@storybook/addon-docs/blocks";
 import * as LocalNavStories from "./local-nav.stories";
-import { PropsTable } from "./props-table";
+import { AutoArgTypesTabs } from "../storybook-common";
 
 <Meta of={LocalNavStories} />
 
@@ -51,4 +51,4 @@ smaller screens.
 
 ## Component API
 
-<PropsTable />
+<AutoArgTypesTabs of={LocalNavStories} />
