+++ stories/masthead/masthead.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta } from "@storybook/addon-docs/blocks";
 import * as MastheadStories from "./masthead.stories";
-import { PropsTable } from "./props-table";
+import { AutoArgTypesTabs } from "../storybook-common";
 
 <Meta of={MastheadStories} />
 
@@ -27,4 +27,4 @@ customised to align with the [Layout](/docs/core-layout--docs) container.
 
 ## Component API
 
-<PropsTable />
+<AutoArgTypesTabs of={MastheadStories} />
