+++ stories/navbar/navbar.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta } from "@storybook/addon-docs/blocks";
 import * as NavbarStories from "./navbar.stories";
-import { PropsTable } from "./props-table";
+import { AutoArgTypesTabs } from "../storybook-common";
 
 <Meta of={NavbarStories} />
 
@@ -124,4 +124,4 @@ The Navbar can be used in conjunction with the Avatar component to display user 
 
 ## Component API
 
-<PropsTable />
+<AutoArgTypesTabs of={NavbarStories} />
