+++ stories/pagination/pagination.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta } from "@storybook/addon-docs/blocks";
 import * as PaginationStories from "./pagination.stories";
-import { PropsTable } from "./props-table";
+import { AutoArgTypesTabs } from "../storybook-common";
 
 <Meta of={PaginationStories} />
 
@@ -42,4 +42,4 @@ Custom page sizes can also be provided.
 
 ## Component API
 
-<PropsTable />
+<AutoArgTypesTabs of={PaginationStories} />
