+++ stories/sidenav/sidenav.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Meta } from "@storybook/addon-docs/blocks";
 import * as SidenavStories from "./sidenav.stories";
-import { PropsTable } from "./props-table";
+import { AutoArgTypesTabs } from "../storybook-common";
 
 <Meta of={SidenavStories} />
 
@@ -22,4 +22,4 @@ The drawer is expanded when a menu item with subitems is hovered or selected. Af
 
 <Canvas of={SidenavStories.SubitemDrawer} />
 
-<PropsTable />
+<AutoArgTypesTabs of={SidenavStories} />
